Create README.md
PyBank and PyPoll Homework assignment using Python. 

The PyBank assignment evaluates a series of financial data broken down by date and Profit/Loss by month. The assignment was to present the total number of months, total net profit/loss, average change and the greatest increase and greatest decrease in profit. 

The PyPoll assignment evaluates election data on fields, voter ID, County and Candidate. The requirement is to present the candidate, the percent of the vote they received and the total number of votes they received and lastly, declare a winner based on the highest number of votes.